Detect when RecordedTV share inaccessible to client
GetBackendVersion now warns user and logs error if path to rec tv does not work

Timer that only was available in windows clients to trigger recording list updates was removed.  Now when timers are set for recording that are,
started immediately swmc won't return until it detects the creation of the new recording.  In that way the 'recordings list' update is guaranteed for
all client platforms.  If after 10s the recording has not started, swmc will time out and return a warning to the client.  Note this works also for instant
recordings and series recordings.